Add SELECTIVE replay mode, record modified retro functions, refactor MDL code.
